Fix bookings form time format in data and select.
Django forms must have changed format. At the time this was written when
a list was provided for a `select` it would replicate the same values as
both values in the tuple. The behaviour changed and some system preset
default formatted value appeared as the first value of the tuple.

When written: [whatever the value] '15:00'
Django change: '3pm.'
This commit: '15:00'