Create new view/template for Games list without category
Adding categories is a bit of a pain and it's not going to happen right now,
as decisions still need to be made and it's probably a good idea to
scrape this data from bgg. We want to focus on being able to display the
list for the purposes of generating the games menu template.